refactor transaction type selection and cache revalidation
Now when opening the transactions page the transaction type is empty
and a menu on the right gets available where the user can choose
the type. When the type is chosen the menu closes and the form
shows up with the selected type. If the user selects a transaction
type from the left menu the transaction type menu is gone and a
button with the transaction type is displayed on the right instead.
If the user clicks on the button, which has the type already set
the form for the type opens up.

Fix cache revalidation after a new transaction is added. After saving
set the api response as `Transactions` before returning to `mutate`.
Before merging the new transaction with the existing list use the
operators `?` and `??` to avoid `undefined` error since a list may
not exist when the added transaction is the first one for the period.
Being the first one means that cache key is not yet available, so
no transactions list exist.